Adjusted the heatmap display scripts to work with the instruction formats passed in through the JSON files. Also moved this into the common directory and added a profiling system to allow the script to be used for other factors
